Remove unusable header stuff
