--- docs/headers/ti/graphscreen.rst
@@ -1,20 +1,0 @@
-.. _graphscreen:
-
-ti/graphscreen.h
-================
-
-.. code-block:: c
-
-    #include <ti/graphscreen.h>
-
-TI made extensive provisions for applications and programs to use the OS's graph screen.
-For example, the Inequalz app does this so that its alternative graphing mode integrates seemlessly.
-
-.. contents:: :local:
-   :depth: 3
-
-API Documentation
------------------
-
-.. doxygenfile:: ti/graphscreen.h
-   :project: CE C/C++ Toolchain
