--- docs/headers/ti/homescreen.rst
@@ -1,20 +1,0 @@
-.. _homescreen:
-
-ti/homescreen.h
-===============
-
-.. code-block:: c
-
-    #include <ti/homescreen.h>
-
-The homescreen is the screen where you can type expressions and get results immediately.
-You can base your UI around the text-only homescreen if you want using these functions.
-
-.. contents:: :local:
-   :depth: 3
-
-API Documentation
------------------
-
-.. doxygenfile:: ti/homescreen.h
-   :project: CE C/C++ Toolchain
