+++ examples/fileioc/factorize_ans/src/main.c
@@ -1,4 +1,4 @@
-#include <ti/homescreen.h>
+#include <ti/screen.h>
 #include <ti/real.h>
 #include <math.h>
 #include <fileioc.h>
