+++ examples/graphx/tilemap_appvar/.gitignore
@@ -1,5 +1,6 @@
 obj/
 bin/
+src/gfx/*.8xv
 src/gfx/*.c
 src/gfx/*.h
 .DS_Store
