+++ examples/hello_world/src/main.c
@@ -1,4 +1,4 @@
-#include <ti/homescreen.h>
+#include <ti/screen.h>
 #include <ti/getcsc.h>
 #include <stdlib.h>
