+++ examples/standalone/math_test/src/main.c
@@ -14,7 +14,7 @@
 
 #ifdef _EZ80
 #   include <debug.h>
-#   include <ti/homescreen.h>
+#   include <ti/screen.h>
 #   include <ti/getcsc.h>
 
 #   if INTERACTIVE
