+++ examples/standalone/random/src/main.c
@@ -1,6 +1,6 @@
 #include <stdlib.h>
 #include <ti/getcsc.h>
-#include <ti/homescreen.h>
+#include <ti/screen.h>
 #include <sys/rtc.h>
 #include <sys/lcd.h>
 #include <sys/util.h>
