+++ examples/standalone/real_time_clock/src/main.c
@@ -1,4 +1,4 @@
-#include <ti/homescreen.h>
+#include <ti/screen.h>
 #include <sys/rtc.h>
 #include <sys/lcd.h>
 #include <sys/util.h>
