+++ examples/standalone/stopwatch/src/main.c
@@ -1,5 +1,5 @@
 #include <ti/getcsc.h>
-#include <ti/homescreen.h>
+#include <ti/screen.h>
 #include <ti/real.h>
 #include <sys/timers.h>
