+++ src/ce/tireal.hpp
@@ -1,0 +1,3 @@
+#include <ti/real>
+#warning <tireal.hpp> is deprecated; use <ti/real> instead
+
